% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{add_clustered_points}
\alias{add_clustered_points}
\title{Add clustered points}
\usage{
add_clustered_points(
  map,
  lon,
  lat,
  cluster_style = NULL,
  marker_style = NULL,
  popup = NULL,
  cluster_options = list(distance = 10, min_distance = 0),
  ...
)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{lon}{numeric: longitudes}

\item{lat}{numeric: latitudes}

\item{cluster_style}{: a style object as returned by \code{\link[=pol_style]{pol_style()}} for the cluster markers, typically with components \code{text} and \code{circle}}

\item{marker_style}{: a style object as returned by \code{\link[=pol_style]{pol_style()}} for single markers (i.e. when a cluster has reduced to a single feature)}

\item{popup}{character: popups to show, one per point}

\item{cluster_options}{list: named list with components \code{distance} and \code{min_distance} (as for \url{https://openlayers.org/en/latest/apidoc/module-ol_source_Cluster-Cluster.html})}

\item{...}{: named vector layer options \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Vector-VectorLayer.html}}
}
\value{
A pol map object
}
\description{
Add clustered points
}
