% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{add_cog}
\alias{add_cog}
\title{Add cloud-optimised geotiff layer}
\usage{
add_cog(map, sources, geotiff_source_options = NULL, ...)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{sources}{: either a character vector of one or more URLs, or a list of \code{pol_geotiff_source} objects}

\item{geotiff_source_options}{list: geotiff source options \url{https://openlayers.org/en/latest/apidoc/module-ol_source_GeoTIFF.html#~GeoTIFFSourceOptions}}

\item{...}{: named webGL tile source options \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_WebGLTile.html}}
}
\value{
A pol map object
}
\description{
Add cloud-optimised geotiff layer
}
