% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgb.R
\name{add_fgb}
\alias{add_fgb}
\title{Add a flatgeobuf vector layer}
\usage{
add_fgb(map, url, file, style = NULL, options = NULL)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{url}{string: URL to the flatgeobuf. Ignored if \code{file} is provided}

\item{file}{string: path to the flatgeobuf file. Note that \code{file} will only work with Shiny, not in standalone htmlwidgets or rmarkdown files}

\item{style}{list: styling as a list of flat style properties: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_flat.html}}

\item{options}{list: vector layer options \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Vector-VectorLayer.html}}
}
\value{
A pol map object
}
\description{
Add a flatgeobuf vector layer
}
\references{
\url{https://flatgeobuf.org/examples/openlayers/}
}
