% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{add_geojson}
\alias{add_geojson}
\title{Add a geoJSON vector layer}
\usage{
add_geojson(
  map,
  data,
  file,
  style = NULL,
  popup = NULL,
  options = NULL,
  data_proj = NULL
)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{data}{string: geoJSON data as a string. Ignored if \code{file} is provided}

\item{file}{string: path to geojson file}

\item{style}{: either a style object as returned by \code{\link[=pol_style]{pol_style()}} or a list of flat style properties: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_flat.html}}

\item{popup}{character: popups to show, one per point}

\item{options}{list: vector layer options \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Vector-VectorLayer.html}}

\item{data_proj}{string: the projection code of the geojson data. If omitted, it will be derived from the data where possible. If it cannot, the data will not be reprojected to the map's projection}
}
\value{
A pol map object
}
\description{
Add a geoJSON vector layer
}
