% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{add_points}
\alias{add_points}
\title{Add points}
\usage{
add_points(map, lon, lat, style = NULL, popup = NULL, ...)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{lon}{numeric: longitudes}

\item{lat}{numeric: latitudes}

\item{style}{: either a style object as returned by \code{\link[=pol_style]{pol_style()}} or a list of flat style properties: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_flat.html}}

\item{popup}{character: popups to show, one per point}

\item{...}{: named vector layer options \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Vector-VectorLayer.html}}
}
\value{
A pol map object
}
\description{
Add points
}
