% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{add_wms_tiles}
\alias{add_wms_tiles}
\title{Add WMS tile layer to a map}
\usage{
add_wms_tiles(
  map,
  url,
  layers,
  wms_params = NULL,
  tile_wms_options = list(),
  options = list()
)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{url}{string: URL to the WMS server}

\item{layers}{character: WMS layer names to include}

\item{wms_params}{list: \code{params} of \url{https://openlayers.org/en/latest/apidoc/module-ol_source_TileWMS-TileWMS.html}. The \code{LAYERS}, \code{TILED}, and some other elements are set automatically}

\item{tile_wms_options}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_source_TileWMS-TileWMS.html}}

\item{options}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Tile-TileLayer.html}}
}
\value{
A pol map object
}
\description{
Add WMS tile layer to a map
}
