% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{add_wmts_from_capabilities}
\alias{add_wmts_from_capabilities}
\title{Add WMTS tile layer to a map from a WMTS server capabilities endpoint}
\usage{
add_wmts_from_capabilities(map, url, layer, wmts_options = list(), ...)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer to}

\item{url}{string: URL to the WMTSCapabilities.xml document}

\item{layer}{string: layer name}

\item{wmts_options}{named list: other WMTS options as per \url{https://openlayers.org/en/latest/apidoc/module-ol_source_WMTS.html#.optionsFromCapabilities}}

\item{...}{: named options as per \url{https://openlayers.org/en/latest/apidoc/module-ol_layer_Tile-TileLayer.html}}
}
\value{
A pol map object
}
\description{
Add WMTS tile layer to a map from a WMTS server capabilities endpoint
}
