% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol.R
\name{pol}
\alias{pol}
\title{Create an openlayers map object}
\usage{
pol(
  view_options = list(),
  control_options = list(zoom = TRUE, rotate = FALSE),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{view_options}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_View.html#~ViewOptions}}

\item{control_options}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_control_defaults.html#~DefaultsOptions}}

\item{width, height}{: Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'}, \code{'auto'}) or a number, which will be coerced to a string and have \code{'px'} appended}

\item{elementId}{string: element ID}
}
\description{
Create an openlayers map object
}
