% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{pol_colourmap}
\alias{pol_colourmap}
\alias{pol_colormap}
\title{Create a colour palette to use with a raster}
\usage{
pol_colourmap(col, range = c(0, 1), breaks, with_nodata = TRUE)

pol_colormap(col, range = c(0, 1), breaks, with_nodata = TRUE)
}
\arguments{
\item{col}{character: vector of hex colour strings}

\item{range}{numeric: data range, noting that by default a layer's values are scaled to the range 0-1}

\item{breaks}{numeric: optional vector of breaks, same length as \code{col}. If not provided, breaks will be equally spaced across \code{range}}

\item{with_nodata}{logical: if \code{TRUE}, the layer's "nodata" (missing) value will be shown as transparent}
}
\value{
An object of class \code{pol_colourmap}
}
\description{
Create a colour palette to use with a raster
}
