% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{pol_proxy}
\alias{pol_proxy}
\title{Modify a pol object inside a Shiny app}
\usage{
pol_proxy(
  outputId,
  session = shiny::getDefaultReactiveDomain(),
  deferUntilFlush = TRUE
)
}
\arguments{
\item{outputId}{string: the ID of the map}

\item{session}{the Shiny session object to which the map belongs; usually the
default value will suffice.}

\item{deferUntilFlush}{logical: if \code{TRUE} then wait until the outputs are next updated to apply changes to the map}
}
\description{
Modify a pol object inside a Shiny app
}
