% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr-package.R
\docType{package}
\name{polr}
\alias{polr}
\alias{polr-package}
\title{\pkg{polr}}
\description{
A minimal set of bindings to openlayers, with a focus on polar mapping applications.
}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

}
