% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_layer}
\alias{remove_layer}
\title{Remove a layer}
\usage{
remove_layer(map, layer_name)
}
\arguments{
\item{map}{pol or pol_proxy: map to remove the layer from}

\item{layer_name}{string: name of the layer. \code{layer_name} must have been provided when adding the layer (e.g. along the lines of \code{add_wms_tiles(..., name = "My Layer")})}
}
\value{
A pol map object
}
\description{
Remove a layer
}
