% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_switcher.R
\name{add_layer_switcher}
\alias{add_layer_switcher}
\title{Add layer switcher}
\usage{
add_layer_switcher(map, target_id)
}
\arguments{
\item{map}{pol or pol_proxy: map to add the layer switcher to}

\item{target_id}{string: ID of the html element to add the switcher to. If missing, the switcher will be placed inside the map pane}
}
\value{
A pol map object
}
\description{
Add layer switcher
}
