% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{pol_style}
\alias{pol_style}
\title{Style}
\usage{
pol_style(
  fill = NULL,
  image = NULL,
  stroke = NULL,
  text = NULL,
  z_index = NULL
)
}
\arguments{
\item{fill}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_Fill-Fill.html}}

\item{image}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_Image-ImageStyle.html} or \url{https://openlayers.org/en/latest/apidoc/module-ol_style_Icon-Icon.html} or \url{https://openlayers.org/en/latest/apidoc/module-ol_style_RegularShape-RegularShape.html}}

\item{stroke}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}}

\item{text}{list: \url{https://openlayers.org/en/latest/apidoc/module-ol_style_Text-Text.html}}

\item{z_index}{integer: z index}
}
\value{
An oject of class \code{pol_style}
}
\description{
Style
}
\examples{
mystyle <- pol_style(stroke = list(color = "green"))

}
\references{
\url{https://openlayers.org/en/latest/apidoc/module-ol_style_Style.html}
}
\seealso{
\code{\link[=add_geojson]{add_geojson()}}
}
