% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr-package.R
\docType{package}
\name{polr}
\alias{polr}
\alias{polr-package}
\title{\pkg{polr}}
\description{
A minimal set of bindings to the OpenLayers javascript library, with a focus on polar mapping applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SCAR/polr/}
  \item Report bugs at \url{https://github.com/SCAR/polr/issues/}
}

}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

Other contributors:
\itemize{
  \item Stefan Kuethe \email{crazycapivara@gmail.com} (Author of https://github.com/crazycapivara/openlayers/, of which polr is a reimplementation and extension (of a sort)) [contributor]
  \item OpenLayers contributors (OpenLayers javascript library) [contributor]
}

}
